Remap of gravity force magnitude
The remap is made so the pull magnitude at the edge of the field can be determined based on the pull magnitude at the surface of the planet. This is done by linear interpolation and gives more control in order to vary  the different radius of planets and their respective fields without making the force at the edge of bigger planet fields negligible.